backend/memory/reranker.py
"""
Cross-Encoder Re-ranker for RAG Retrieval

OPTIMIZATION 4: Re-rank retrieved memories using cross-encoder for better precision.

This module provides intelligent re-ranking of vector search results using
cross-encoder/stsb-roberta-large, trained on the Semantic Textual Similarity 
Benchmark for precise semantic similarity scoring.

Strategy:
1. Retrieve more candidates than needed (e.g., top_k=10)
2. Use STS-B cross-encoder to score each candidate's relevance to query
3. Re-rank by relevance scores
4. Return only top N results

Expected gain: +15-25% precision improvement
No API costs, no rate limits!
"""
from __future__ import annotations

import logging
import hashlib
from typing import List, Dict, Optional

from memory.vector_store import RetrievalResult


class LLMReranker:
    """
    Re-ranks retrieval results using cross-encoder relevance scoring.
    
    Uses a HuggingFace cross-encoder model to evaluate how relevant each 
    retrieved memory is to the query, then re-ranks by relevance score.
    
    Cross-encoders are specifically trained for this task and provide:
    - Better accuracy than vector similarity alone
    - No rate limits (runs locally)
    - Faster inference (no network latency)
    - Zero cost
    """
    
    def __init__(
        self,
        model_name: str = "cross-encoder/stsb-roberta-large",  # STS-B trained model (~1.3GB) - semantic similarity
        enable_reranking: bool = True,
        use_cache: bool = True,  # Cache scores to avoid redundant computation
        vector_weight: float = None,  # Weight for vector score (0-1), None = read from env
        cross_encoder_weight: float = None,  # Weight for CE score (0-1), None = read from env
        score_threshold: float = None,  # Minimum score threshold, None = read from env
        normalization: str = None  # Normalization method, None = read from env
    ):
        """
        Initialize the cross-encoder re-ranker.
        
        Args:
            model_name: HuggingFace cross-encoder model 
                       (default: cross-encoder/stsb-roberta-large, trained on Semantic 
                        Textual Similarity Benchmark for semantic similarity scoring)
            enable_reranking: Whether to enable re-ranking (can disable for testing)
            use_cache: Whether to cache relevance scores
            vector_weight: Weight for original vector similarity score
            cross_encoder_weight: Weight for cross-encoder relevance score
            score_threshold: Minimum combined score to include result
            normalization: Score normalization method ("minmax" or "sigmoid")
        """
        self.logger = logging.getLogger(self.__class__.__name__)
        self.model_name = model_name
        self.enable_reranking = enable_reranking
        self.use_cache = use_cache
        
        # Hybrid scoring configuration (can be overridden via environment)
        import os
        self.vector_weight = vector_weight if vector_weight is not None else float(os.getenv('RERANKER_VECTOR_WEIGHT', '0.7'))  # TUNED: 0.7 vector
        self.cross_encoder_weight = cross_encoder_weight if cross_encoder_weight is not None else float(os.getenv('RERANKER_CE_WEIGHT', '0.3'))  # TUNED: 0.3 CE
        self.score_threshold = score_threshold if score_threshold is not None else float(os.getenv('RERANKER_THRESHOLD', '0.0'))
        self.normalization = normalization if normalization is not None else os.getenv('RERANKER_NORMALIZATION', 'minmax')
        
        # Simple in-memory cache for scores
        self._score_cache: Dict[str, float] = {}
        
        if not self.enable_reranking:
            self.logger.info("Cross-encoder re-ranking disabled")
            return
        
        # Initialize cross-encoder model
        # PYTHON 3.13 COMPATIBILITY: Use ONNX Runtime instead of PyTorch-based sentence-transformers
        try:
            import sys
            python_version = f"{sys.version_info.major}.{sys.version_info.minor}"
            
            # For Python 3.13+, use ONNX Runtime (no torch dependency)
            if sys.version_info >= (3, 13):
                self.logger.info(f"Python {python_version} detected - using lightweight scoring fallback")
                self.logger.warning("Cross-encoder not available in Python 3.13 (torch compatibility)")
                self.logger.info("Using vector scores only - consider Python 3.11 for full cross-encoder support")
                self.enable_reranking = False
                self.model = None
                return
            
            # For Python < 3.13, use sentence-transformers CrossEncoder
            from sentence_transformers import CrossEncoder
            
            self.model = CrossEncoder(model_name)
            self.logger.info(f"✅ Cross-encoder re-ranker initialized: {model_name}")
            self.logger.info(f"   Hybrid scoring: Vector={self.vector_weight:.2f}, CE={self.cross_encoder_weight:.2f}")
            self.logger.info(f"   Threshold: {self.score_threshold:.2f}, Normalization: {self.normalization}")
                
        except Exception as e:
            self.logger.error(f"Failed to initialize cross-encoder re-ranker: {e}")
            self.logger.warning("Disabling re-ranking - falling back to vector scores")
            self.enable_reranking = False
    
    def rerank(
        self,
        query: str,
        results: List[RetrievalResult],
        top_k: int = 5
    ) -> List[RetrievalResult]:
        """
        Re-rank results using hybrid scoring (vector + cross-encoder).
        
        Args:
            query: The search query
            results: List of retrieval results to re-rank
            top_k: Number of top results to return after re-ranking
            
        Returns:
            Re-ranked list of results (top_k items)
        """
        if not self.enable_reranking or not results:
            return results[:top_k]
        
        try:
            # Score each result with hybrid approach
            scored_results = []
            for result in results:
                # Get cross-encoder score
                ce_score = self._score_relevance(query, result.text)
                
                # Hybrid scoring: combine vector similarity + cross-encoder relevance
                vector_score = result.score  # Original similarity score from vector search
                combined_score = (self.vector_weight * vector_score) + (self.cross_encoder_weight * ce_score)
                
                # Apply threshold filter
                if combined_score >= self.score_threshold:
                    # Create new result with combined score
                    scored_result = RetrievalResult(
                        id=result.id,
                        text=result.text,
                        score=combined_score,  # Hybrid score
                        metadata=result.metadata,
                        rank=result.rank,
                        vector_score=result.vector_score or result.score,
                        lexical_score=result.lexical_score,
                    )
                    scored_results.append(scored_result)
            
            # Sort by combined score (descending)
            scored_results.sort(key=lambda x: x.score, reverse=True)
            
            self.logger.debug(f"Re-ranked {len(results)} → {len(scored_results)} results (after threshold), returning top {top_k}")
            
            # Return top_k
            return scored_results[:top_k]
            
        except Exception as e:
            self.logger.error(f"Re-ranking failed: {e}")
            # Fallback to original results
            return results[:top_k]
    
    def _score_relevance(self, query: str, document: str) -> float:
        """
        Score how relevant a document is to a query using cross-encoder.
        
        Args:
            query: The search query
            document: The retrieved document text
            
        Returns:
            Relevance score between 0.0 and 1.0
        """
        # Check cache first
        if self.use_cache:
            cache_key = hashlib.md5(f"{query}::{document}".encode()).hexdigest()
            if cache_key in self._score_cache:
                self.logger.debug(f"Cache hit for relevance score")
                return self._score_cache[cache_key]
        
        try:
            # Cross-encoder directly scores the (query, document) pair
            # Returns a single score (higher = more relevant)
            raw_score = self.model.predict([(query, document)])[0]
            
            # Normalize based on strategy
            if self.normalization == "sigmoid":
                # Sigmoid normalization (for logit-style scores)
                import math
                score = 1 / (1 + math.exp(-float(raw_score)))
            
            elif self.normalization == "minmax":
                # Min-max normalization (MS-MARCO scores typically range -5 to +15)
                score = float(raw_score)
                min_score = -5.0
                max_score = 15.0
                score = (score - min_score) / (max_score - min_score)
            
            else:  # raw
                # Use raw scores (may need manual interpretation)
                score = float(raw_score)
            
            # Clamp to [0, 1]
            score = max(0.0, min(1.0, score))
            
            # Cache the score
            if self.use_cache:
                self._score_cache[cache_key] = score
            
            self.logger.debug(f"CE score: {score:.3f} (raw: {raw_score:.2f}, norm: {self.normalization})")
            return score
            
        except Exception as e:
            self.logger.error(f"Scoring failed: {e}")
            # Fallback to neutral score
            return 0.5


# Global singleton
_reranker: Optional[LLMReranker] = None


def get_reranker(reset: bool = False, **kwargs) -> LLMReranker:
    """
    Get or create global cross-encoder re-ranker instance.
    
    Args:
        reset: Whether to reset the singleton
        **kwargs: Arguments for LLMReranker constructor
        
    Returns:
        LLMReranker instance
    """
    global _reranker
    
    if reset or _reranker is None:
        _reranker = LLMReranker(**kwargs)
    
    return _reranker
