backend/requirements.txt
# ===========================================
# ATLAS - AI-Powered Misinformation Fighter
# Updated Requirements File
# ===========================================

# Core Web Framework
quart==0.20.0
quart-cors==0.8.0
hypercorn==0.17.3
werkzeug==3.1.3

# HTTP & Networking
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
h2==4.3.0
certifi==2025.10.5
urllib3==2.3.0
idna==3.11

# Web Scraping & Content Extraction
beautifulsoup4==4.14.2
fake-useragent==2.2.0
trafilatura==2.0.0
lxml==5.4.0
soupsieve==2.8
playwright==1.55.0

# AI & ML Models
groq==0.33.0
# Shared constraint: fastembed (<1.0) and transformers (>=0.34)
huggingface-hub==0.36.0
fastembed==0.7.3

# Async & File Operations
aiofiles==25.1.0
aiosqlite==0.21.0
aiohttp==3.11.11
anyio==4.11.0

# Configuration & Environment
python-dotenv==1.2.1
pydantic==2.12.4
pydantic-core==2.41.5
tenacity==9.1.2

# Rate Limiting & Security
limits==5.6.0
deprecated==1.3.1
wrapt==2.0.0

# Date & Time Utilities
python-dateutil==2.9.0.post0

# CLI & Progress
tqdm==4.67.1
click==8.3.0
colorama==0.4.6
shellingham==1.5.4
typer-slim==0.20.0

# Data Handling
pyyaml==6.0.3
fsspec==2025.10.0
filelock==3.20.0

# OCR & Image Processing (EasyOCR - no Tesseract installation needed!)
easyocr==1.7.2
pillow==11.0.0
torch==2.9.0
torchvision==0.24.0
opencv-python-headless==4.12.0.88
scipy==1.16.3
scikit-image==0.25.2
numpy==2.2.6
rank-bm25==0.2.2

# Web Server Protocols
wsproto==1.2.0
hpack==4.1.0
hyperframe==6.1.0
priority==2.0.0

# Type Checking & Utilities
typing-extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0
packaging==25.0
sniffio==1.3.1

# Template Engine
jinja2==3.1.6
markupsafe==3.0.3

# Async Support
greenlet==3.2.4
pyee==13.0.0

# Distributed Systems
distro==1.9.0
hf-xet==1.2.0

# Flask (for compatibility)
flask==3.1.2
blinker==1.9.0
itsdangerous==2.2.0

# Character Encoding
charset-normalizer==3.4.4

# ===========================================
# Optional Dependencies (Install as needed)
# ===========================================

# Text processing and NLP (Optional)
# textblob==0.19.0
# sumy==0.11.0

# Data visualization (Optional)
# tabulate==0.9.0
# wordcloud==1.9.4
# matplotlib==3.10.7

# NumPy (Optional - requires C compiler on Windows)
# numpy==1.26.4

# ML libraries (Optional - heavy dependencies)
# spacy>=3.4,<3.8
# transformers>=4.21,<4.30
# torch>=1.13,<2.2

# Database ORM (Optional)
# SQLAlchemy==2.0.23

# Development tools (Optional)
# pytest==8.4.2
# black==25.9.0
# flake8==7.3.0
# mypy==1.18.2
# isort==6.0.1

# Additional utilities (Optional)
# pillow==12.0.0